+++ src/main.tsx
@@ -5,7 +5,7 @@ import { createRoot } from 'react-dom/client';
 import { BrowserRouter } from "react-router";
 import App from './App.tsx';
 
-import AuthProvider from "./components/AuthProvider.tsx"
+import AuthProvider from "./context//AuthProvider.tsx"
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
